Dev branch sailesh

See merge request csce_361/spring2020/30pair17!3

+++ src/main/java/edu/unl/cse/csce361/petting_zoo/model/AnimalBuilder.java
@@ -76,8 +76,7 @@ public class AnimalBuilder extends Observable {
     }
 
     public Animal build() {
-        //where price = pricePerKg*MassInKg
-        return new AnimalEntity(name, type, sex, massInKg, (pricePerKg*massInKg), species,
+        return new AnimalEntity(name, type, sex, massInKg, ((double)pricePerKg*(double)massInKg), species,
                 reactionToPetting, reactionToFeeding, reactionToWatching);
     }
